+++ src/RcppExports.cpp
@@ -700,6 +700,200 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// impl_tree_build
+SEXP impl_tree_build(SEXP x);
+RcppExport SEXP _pjrt_impl_tree_build(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_build(x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_flatten
+Rcpp::List impl_tree_flatten(SEXP x);
+RcppExport SEXP _pjrt_impl_tree_flatten(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_flatten(x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_build_flatten
+Rcpp::List impl_tree_build_flatten(SEXP x);
+RcppExport SEXP _pjrt_impl_tree_build_flatten(SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_build_flatten(x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_unflatten
+SEXP impl_tree_unflatten(SEXP tree, Rcpp::List x);
+RcppExport SEXP _pjrt_impl_tree_unflatten(SEXP treeSEXP, SEXP xSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    Rcpp::traits::input_parameter< Rcpp::List >::type x(xSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_unflatten(tree, x));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_size
+int impl_tree_size(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_size(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_size(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_equal
+bool impl_tree_equal(SEXP a, SEXP b);
+RcppExport SEXP _pjrt_impl_tree_equal(SEXP aSEXP, SEXP bSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type a(aSEXP);
+    Rcpp::traits::input_parameter< SEXP >::type b(bSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_equal(a, b));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_kind
+std::string impl_tree_kind(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_kind(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_kind(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_names
+SEXP impl_tree_names(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_names(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_names(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_child_kinds
+Rcpp::CharacterVector impl_tree_child_kinds(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_child_kinds(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_child_kinds(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_child_sizes
+Rcpp::IntegerVector impl_tree_child_sizes(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_child_sizes(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_child_sizes(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_flat_names
+Rcpp::CharacterVector impl_tree_flat_names(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_flat_names(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_flat_names(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_path
+std::string impl_tree_path(SEXP tree, int i);
+RcppExport SEXP _pjrt_impl_tree_path(SEXP treeSEXP, SEXP iSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    Rcpp::traits::input_parameter< int >::type i(iSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_path(tree, i));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_filter_by_names
+SEXP impl_tree_filter_by_names(SEXP tree, Rcpp::CharacterVector names);
+RcppExport SEXP _pjrt_impl_tree_filter_by_names(SEXP treeSEXP, SEXP namesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    Rcpp::traits::input_parameter< Rcpp::CharacterVector >::type names(namesSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_filter_by_names(tree, names));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_concat
+SEXP impl_tree_concat(Rcpp::List children, SEXP names);
+RcppExport SEXP _pjrt_impl_tree_concat(SEXP childrenSEXP, SEXP namesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::List >::type children(childrenSEXP);
+    Rcpp::traits::input_parameter< SEXP >::type names(namesSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_concat(children, names));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_mask_from_names
+Rcpp::LogicalVector impl_tree_mask_from_names(SEXP tree, Rcpp::CharacterVector names);
+RcppExport SEXP _pjrt_impl_tree_mask_from_names(SEXP treeSEXP, SEXP namesSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    Rcpp::traits::input_parameter< Rcpp::CharacterVector >::type names(namesSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_mask_from_names(tree, names));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_repr
+std::string impl_tree_repr(SEXP tree);
+RcppExport SEXP _pjrt_impl_tree_repr(SEXP treeSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type tree(treeSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_repr(tree));
+    return rcpp_result_gen;
+END_RCPP
+}
+// impl_tree_diff
+SEXP impl_tree_diff(SEXP a, SEXP b);
+RcppExport SEXP _pjrt_impl_tree_diff(SEXP aSEXP, SEXP bSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type a(aSEXP);
+    Rcpp::traits::input_parameter< SEXP >::type b(bSEXP);
+    rcpp_result_gen = Rcpp::wrap(impl_tree_diff(a, b));
+    return rcpp_result_gen;
+END_RCPP
+}
 // impl_test_xptr_prot
 SEXP impl_test_xptr_prot(SEXP x);
 RcppExport SEXP _pjrt_impl_test_xptr_prot(SEXP xSEXP) {
@@ -773,6 +967,23 @@ static const R_CallMethodDef CallEntries[] = {
     {"_pjrt_get_orgqr_handler_cuda", (DL_FUNC) &_pjrt_get_orgqr_handler_cuda, 0},
     {"_pjrt_get_svd_handler", (DL_FUNC) &_pjrt_get_svd_handler, 0},
     {"_pjrt_get_svd_handler_cuda", (DL_FUNC) &_pjrt_get_svd_handler_cuda, 0},
+    {"_pjrt_impl_tree_build", (DL_FUNC) &_pjrt_impl_tree_build, 1},
+    {"_pjrt_impl_tree_flatten", (DL_FUNC) &_pjrt_impl_tree_flatten, 1},
+    {"_pjrt_impl_tree_build_flatten", (DL_FUNC) &_pjrt_impl_tree_build_flatten, 1},
+    {"_pjrt_impl_tree_unflatten", (DL_FUNC) &_pjrt_impl_tree_unflatten, 2},
+    {"_pjrt_impl_tree_size", (DL_FUNC) &_pjrt_impl_tree_size, 1},
+    {"_pjrt_impl_tree_equal", (DL_FUNC) &_pjrt_impl_tree_equal, 2},
+    {"_pjrt_impl_tree_kind", (DL_FUNC) &_pjrt_impl_tree_kind, 1},
+    {"_pjrt_impl_tree_names", (DL_FUNC) &_pjrt_impl_tree_names, 1},
+    {"_pjrt_impl_tree_child_kinds", (DL_FUNC) &_pjrt_impl_tree_child_kinds, 1},
+    {"_pjrt_impl_tree_child_sizes", (DL_FUNC) &_pjrt_impl_tree_child_sizes, 1},
+    {"_pjrt_impl_tree_flat_names", (DL_FUNC) &_pjrt_impl_tree_flat_names, 1},
+    {"_pjrt_impl_tree_path", (DL_FUNC) &_pjrt_impl_tree_path, 2},
+    {"_pjrt_impl_tree_filter_by_names", (DL_FUNC) &_pjrt_impl_tree_filter_by_names, 2},
+    {"_pjrt_impl_tree_concat", (DL_FUNC) &_pjrt_impl_tree_concat, 2},
+    {"_pjrt_impl_tree_mask_from_names", (DL_FUNC) &_pjrt_impl_tree_mask_from_names, 2},
+    {"_pjrt_impl_tree_repr", (DL_FUNC) &_pjrt_impl_tree_repr, 1},
+    {"_pjrt_impl_tree_diff", (DL_FUNC) &_pjrt_impl_tree_diff, 2},
     {"_pjrt_impl_test_xptr_prot", (DL_FUNC) &_pjrt_impl_test_xptr_prot, 1},
     {NULL, NULL, 0}
 };